Adding check to shared library

+++ Util/BuildTools/BuildUtilsDocker.sh
@@ -41,7 +41,9 @@ cmake -G "Ninja" \
     ..
 
 # copy the shared object 'libfbxsdk.so' to 'dist' folder
-cp "${FBX2OBJ_DEP_FOLDER}/lib/gcc/x64/release/libfbxsdk.so" "${FBX2OBJ_DIST}"
+if [ ! -f "${FBX2OBJ_DIST}/libfbxsdk.so" ]; then
+  cp "${FBX2OBJ_DEP_FOLDER}/lib/gcc/x64/release/libfbxsdk.so" "${FBX2OBJ_DIST}"
+fi
 
 set +e
